Update README.md
The goal of our project is to create a tool that ensures the safety and hygiene of frontline workers in retail stores. Our project aims to use Computer Vision technology to classify whether the employee/customer is attempting to enter the store premises without a mask. If the algorithm recognizes that an individual is not wearing a mask, then a mask is automatically dispensed by the machine and the algorithm again checks if the mask is worn by the person. Especially, for employees, the program uses a rewarding system that awards them for following adequate guidelines to stop the spread of the pandemic.